feat(lazy_disk_cache): add generic `DiskBackedStore`
`DiskBackedStore` can be mostly used as a drop in replacement for
`pc2img.image_store.DiskBackedImageStore`. The store is implemented
in such a way that it will greedly add any existing files from the
cache dir. The interaction between `LazyDiskCacheKw` and
`LazyDiskCacheConfig` has been strengthend by adding back-and-forth
functions between the two.